+++ CHANGELOG.md
@@ -7,7 +7,8 @@
   - `#154`: As dev, I want to change all `console.log` to `console.info` so that i can standardize the codes.
 - **Test**
   - `#160`: As test, I want to jest test and eslint test for all non-master branch commit
-  - `#164`: As test, I want to standardize test data set for `true`.
+  - `#164`: As test, I want to standardize test data set for `true` (`common-data-sets.js/TrueDataSet`).
+  - `#165`: As test, I want to standardize test data set for `false` (`common-data-sets.js/FalseDataSet`).
 
 ## `0.12.0` - 2021-Jul-27